chore(exchange): clean up and format code
